+++ CHANGELOG.md
@@ -2,9 +2,9 @@
 
 ## v0.0.12 (unreleased)
 
- * Add `ListFiles` and `ListJobs` to Hist API
+ * Add `SubmitJobs`, `ListFiles`, and `ListJobs` to Hist API
  * Add some JSON marshallers
- * `dbn-go-hist` now has `jobs` and `files` subcommands
+ * `dbn-go-hist` now has `submit`, `jobs` and `files` subcommands
  * `dbn-go-hist` now supports `--file` to supply lists of tickers
 
 ## v0.0.11 (2024-06-03)